requirements.txt
python-decouple
flask
langchain
ctransformers[cuda]
pywebview


